requirements.txt
PyQt5~=5.15.1
matplotlib~=3.3.2
PySide2~=5.15.1
mysql-connector-python~=8.0.21
XlsxWriter~=1.3.7